Fix build failure of SPL because libc-dev is missing (LP: #1536299)
DKMS SPL build failure occurs on a clean machine because the configure
script attempts to build a small test C program which fails to link
because libc-dev is not installed at that point even though gcc is
installed.

configure: error:
        *** Please make sure the kmod spl devel <kernel> package for your
        *** distribution is installed then try again.  If that fails you
        *** can specify the location of the spl objects with the
        *** '--with-spl-obj=PATH' option.

The most straight forward fix is to state that libc-dev is a dependency
so that it is installed first before SPL is built.  Most DKMS builds
just require gcc and build directly against the kernel, but for SPL
we have this pre-build configure step which needs libc-dev.

Signed-off-by: Colin Ian King <colin.king@canonical.com>